feature: React Aria LinkコンポーネントとtypedRoutesを統合
Next.jsのtypedRoutesを有効化し、React Aria ComponentsのLinkを実装。
RouterAdapterを改善してModule Augmentationで型安全性を確保。

主な変更:
- next.config.ts: typedRoutesを有効化
- RouterAdapter: _components/配下に移動、navigate関数を明示的に定義
- Link: React Aria ComponentsベースのLinkコンポーネントを新規追加
- Module Augmentation: LinkConfigでhref型をカスタマイズ可能に
