+++ apps/web/next.config.ts
@@ -5,6 +5,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
 	reactCompiler: true,
 	cacheComponents: true,
+	typedRoutes: true,
 };
 
 export default withSentryConfig(nextConfig, {
