+++ apps/web/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import { ThemeProvider } from "next-themes";
-import { RouterAdapter } from "./router-adapter";
+import { RouterAdapter } from "./_components/router-adapter";
 import "./globals.css";
 
 const geistSans = Geist({
