+++ packages/react-components/src/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./button";
+export * from "./link";
